Re-organize test structure
